Correct the ur-off-by-one that caused all these problems
I had an off-by-one error that compounded into multiple horrifying
problems: By inflating the "first" range to the right, we ended up
reading too far (resulting in uninitialized data being possible to
read sometimes). Oops!

That caused several bugs that compensated for the original
problem (and was caused by them in turn), and uuuusually evened
out.

But this is the fact: The "end()" method one past the last index that
can be read (and was wrongly documented to return the index of the
last element that can be read), which means it's exactly a match for
the End field in Range structs. We do not have to add anything to it.