Completely re-do Range FIFO/LIFO iteration, get it right
This is so extremely embarrassing, I have no words: I got FIFO and
LIFO mixed up - mixed up enough that all my test cases (including
examples & docs!) were broken, in a way that masked the off-by-one
error that I'm fixing on this branch. So:

* Get LIFO and FIFO straightened out: FIFO is oldest-first, the
  "normal" traversal of a ring. LIFO is youngest-first.

* Restructure how we walk the index Ranges, by making an indexes
  structure that makes Scanner implementation easy & straightforward
  in both directions, with no special cases.

This is (no kidding!) a backwards-incompatible change.